+++ R/orbital_speed.R
@@ -13,7 +13,9 @@
 #'
 #' @examples
 #' # Compute orbital speed for varying Hs
-#' S <- t(sapply(1:10, \(h) jonswap(h)$spec))
+#' S <- t(sapply(1:10, function(h) {
+#'   jonswap(h)$spec
+#' }))
 #' orb_speeds <- compute_orbital_speeds(S, rscd_freq, depth = 100, z = 10)
 #' plot(1:10, orb_speeds[, 1],
 #'   type = "l",
@@ -22,7 +24,13 @@
 #'   ylab = "Orbital speed RMS (m/s)"
 #' )
 #' lines(1:10, orb_speeds[, 2], type = "l", col = "red")
-compute_orbital_speeds <- function(spec, freq, z = 0, depth = Inf, output_speeds = FALSE) {
+compute_orbital_speeds <- function(
+  spec,
+  freq,
+  z = 0,
+  depth = Inf,
+  output_speeds = FALSE
+) {
   # z: distance above sea floor
 
   dims <- dim(spec)
@@ -34,20 +42,23 @@ compute_orbital_speeds <- function(spec, freq, z = 0, depth = Inf, output_speeds
   }
 
   stopifnot(n_freq == length(freq))
-  stopifnot(all(z >= 0))
-  stopifnot(all(z <= depth))
+  stopifnot(z >= 0)
+  stopifnot(z <= depth)
 
   # Compute k efficiently when depth is discretized
   k <- outer(
-    freq, unique(depth),
-    Vectorize(resourcecode::dispersion, vectorize.args = c("frequencies", "depth"))
+    freq,
+    unique(depth),
+    Vectorize(
+      resourcecode::dispersion,
+      vectorize.args = c("frequencies", "depth")
+    )
   )
   mat_k <- t(k[, match(depth, unique(depth))])
 
   mat_d <- matrix(depth, nrow = n_time, ncol = n_freq)
   mat_depth <- matrix(z, nrow = n_time, ncol = n_freq)
   mat_freq <- matrix(freq, nrow = n_time, ncol = n_freq, byrow = TRUE)
-
 
   csh1 <- cosh(mat_k * mat_depth)
   ssh1 <- sinh(mat_k * mat_depth)