+++ tests/testthat/test-test_utils.R
@@ -20,7 +20,10 @@ test_that("Node selection works", {
   node_test <- 134938
   spec_test <- 2124
   expect_equal(closest_point_field(point_test)$point, node_test)
-  expect_equal(closest_point_field(point_test[1], point_test[2])$point, node_test)
+  expect_equal(
+    closest_point_field(point_test[1], point_test[2])$point,
+    node_test
+  )
   expect_equal(closest_point_spec(point_test)$point, 2124)
   expect_equal(closest_point_spec(point_test[1], point_test[2])$point, 2124)
 })
@@ -39,4 +42,56 @@ test_that("JONSWAP computation works", {
   expect_snapshot_output(jonswap())
   expect_snapshot_output(jonswap(fmax = 0.95, df = 0.003))
   expect_error(jonswap(hs = 4, tp = 15, fmax = 0.95))
+  expect_error(jonswap(gam = 0.5))
+})
+
+test_that("Directional means are accurately computed", {
+  expect_error(mean_direction())
+  expect_equal(mean_direction(rep(0, 10)), 0)
+  expect_equal(mean_direction(rep(0, 10)), mean_direction(rep(360, 10)))
+})
+
+test_that("Weighted directional means are accurately computed", {
+  expect_warning(mean_direction(c(0, 90), weights = c(-1, 1)))
+  expect_error(mean_direction(c(0, 90), weights = c(1)))
+  expect_equal(mean_direction(c(0, 90), weights = c(50, 50)), 45)
+  expect_equal(mean_direction(c(0, 90), weights = c(0, 1)), 90)
+})
+
+test_that("Directional binning works", {
+  expect_equal(
+    cut_directions(c(10, 80, 170, 280), n_bins = 4),
+    factor(c("N", "E", "S", "W"), levels = c("N", "E", "S", "W"))
+  )
+  expect_equal(
+    cut_directions(c(10, 80, 170, 280), n_bins = 8),
+    factor(
+      c("N", "E", "S", "W"),
+      levels = c("N", "NE", "E", "SE", "S", "SW", "W", "NW")
+    )
+  )
+  expect_equal(
+    cut_directions(c(10, 80, 170, 280), n_bins = 16),
+    factor(
+      c("N", "E", "S", "W"),
+      levels = c(
+        "N",
+        "NNE",
+        "NE",
+        "ENE",
+        "E",
+        "ESE",
+        "SE",
+        "SSE",
+        "S",
+        "SSW",
+        "SW",
+        "WSW",
+        "W",
+        "WNW",
+        "NW",
+        "NNW"
+      )
+    )
+  )
 })