+++ NEWS.md
@@ -1,3 +1,14 @@
+# resourcecode 0.3.0
+
+# resourcecode 0.2.2
+
+- `rscd_mapplot()`: add Resourcecode caption.
+- New helpers:
+  - `mean_direction()` to compute (weighted) mean values of directional variables,
+  - `cut_directions()` to bin a vector of directions into sectors, centered North;
+  - `cut_seasons()` to compute the season from a datetime, with different definitions.
+ - Fixed #7 (thanks @paalves).
+
 # resourcecode 0.2.1
 
 - Move out the data from the package to `{resourcecodedata}` package to be compliant with